refactor: HandlerMapping 공통 로직 인터페이스 분리
