+++ app/src/main/java/com/techcourse/HandlerMapping.java
@@ -1,0 +1,8 @@
+package com.techcourse;
+
+import jakarta.servlet.http.HttpServletRequest;
+
+public interface HandlerMapping {
+
+    Object getHandler(final HttpServletRequest request);
+}
